+++ src/i_video.c
@@ -2001,6 +2001,8 @@ void I_InitGraphics(void)
     SDL_SetHintWithPriority(SDL_HINT_VIDEO_MINIMIZE_ON_FOCUS_LOSS, "1", SDL_HINT_OVERRIDE);
 #endif
 
+    SDL_SetHintWithPriority(SDL_HINT_RENDER_BATCHING, "0", SDL_HINT_OVERRIDE);
+
     if (vid_fullscreen)
         SDL_ShowCursor(false);