+++ src/z_zone.h
@@ -59,7 +59,7 @@ enum
 #define PU_PURGELEVEL    PU_CACHE    // First purgeable tag's level
 
 void *Z_Malloc(size_t size, unsigned char tag, void **user) ALLOCATTR(1);
-void *Z_Calloc(size_t n1, size_t n2, unsigned char tag, void **user) ALLOCSATTR(1, 2);
+void *Z_Calloc(size_t size1, size_t size2, unsigned char tag, void **user) ALLOCSATTR(1, 2);
 char *Z_StringDuplicate(const char *, unsigned char tag, void **user);
 void Z_Free(void *ptr);
 void Z_FreeTags(unsigned char lowtag, unsigned char hightag);