+++ releasenotes.md
@@ -17,6 +17,7 @@
   * Hotplugging controllers now always works.
   * Pressing the left thumbstick of a controller is no longer bound to the `+alwaysrun` action by default.
   * Controllers no longer rumble if the mouse or keyboard are being used by the player instead.
+* The aspect ratio of the display when in widescreen mode can now be forced by changing the new `vid_aspectratio` CVAR from its default `auto` to `16:9`, `16:10`, `21:9` or `32:9`.
 * Pressing the <kbd><b>ENTER</b></kbd> key to close the help screen no longer causes the previous player message to be displayed.
 * The `save` and `load` CCMDs can now be used to save and load games by specifying numbers `1` to `8`.
 * The player’s currently equipped weapon is now translucent when they have a partial invisibility power-up and the `r_textures` CVAR is `off`.
@@ -25,7 +26,6 @@
 * A bug is fixed whereby the ammo the player had for their equipped weapon wasn’t displayed in the widescreen HUD while they were dead.
 * The power-up bar is now displayed in the alternate widescreen HUD when the player enters the `IDBEHOLDL` cheat.
 * Minor improvements have been made to the highlight effect in the menu in some instances.
-* The aspect ratio of the display when in widescreen mode can now be forced by changing the new `vid_aspectratio` CVAR. This CVAR is `auto` by default and can be set to `16:9`, `16:10`, `21:9` or `32:9`.
 
 ![](https://github.com/bradharding/www.doomretro.com/raw/master/wiki/bigdivider.png)